ansible: Create QEMU VMs
Automate the creation of QEMU virtual machines, including one control
node and two worker nodes, for the Kubernetes cluster setup.

This involves establishing two host groups within the Ansible inventory:
a general group for all inventory hosts and a specific qemu group
dedicated to the QEMU virtual machines.

The process of creating the VMs is delegated to a `vm_host` node, which
is currently set to localhost.

Signed-off-by: Dimitris Poulopoulos <dimitris.a.poulopoulos@gmail.com>