ansible/qemu-vm: Check if QCOW2 image exists
Running the task that creates a QCOW2 image once successfully creates a
new QCOW2 image. But this is not idempotent. Running it a second time
will create a new QCOW2 image with the same name, deleting the old one.

Check if the file that this commands tries to create is there, and
create it only if it does not exist. To this end, use the `creates`
parameter of the `ansible.builtin.command` module. From the docs:

`creates`: A filename or (since 2.0) glob pattern. If a matching file
           already exists, this step will not run.

Signed-off-by: Dimitris Poulopoulos <dimitris.a.poulopoulos@gmail.com>