ansible/qemu-vm: Handle virt module's non-idempotency
The `community.libvirt.virt` Ansible module offers several commands to
define, start, shutdown, etc. domains (i.e., VMs). These commands are
not idempotent, something that they also state in the documentation:

https://docs.ansible.com/ansible/latest/collections/community/libvirt/virt_module.html#parameter-command.

Running the `define` command once will define the domain successfully.
However, running it a second time will define it again and change
certain properties of the domain. For example, since we don't declare
the MAC address of the VM in the XML template, the first time we run the
`define` command, a random MAC address will be asigned to the VM.
The second time we run the task, since it is not idempotent, a new
random MAC address will be assigned and Ansible will show that the VM
was changed.

See the following GItHub issue:

https://github.com/ansible/ansible/issues/27905

Use a new `state.ini` file to keep a record of prperties we want to
assing to the VM once. To make this work, there is a task that looks
for the variable `mac_address` into the hostvars map. If it does not
find one, it creates a random MAC address using the
`community.general.random_mac` module. Then, it records the new random
MAC address as a fact. The next step, writes the MAC address in the
newly created `state.ini` file and reloads the inventory. The second
time this play will run, the former step will find a MAC address into
the hostvars map and re-use it.

Signed-off-by: Dimitris Poulopoulos <dimitris.a.poulopoulos@gmail.com>