ansible/pxe-server: Add GRUB configuration file
By dafault, the GRUB config defines several entries, such as expert
installation and automated installation. It won't start an automated
installation by default, and even if we choose it manually, it won't
fetch the preseed file.

Create a new GRUB config file that sets the automated installation
option as the default one, make it start after 5 seconds of inactivity,
and tell GRUB how to fetch the preseed file.

Signed-off-by: Dimitris Poulopoulos <dimitris.a.poulopoulos@gmail.com>