ansible/pxe-server: Simplify Debian version check
Simplify the rules that define whether download the Debian installer and
extract it.

Instead of getting the version of the installer via the `version.info`
file and comparing it with the user defined one, hash the downloaded
tarball, if present, and compare it before downloading it using
Ansible's builtin modules.

Extract it only if the file `version.info` is not present, since this is
the last file that gets extracted from the package. This way, if the
extraction process terminates unsuccessfully, the script remains
idempotent.

Signed-off-by: Dimitris Poulopoulos <dimitris.a.poulopoulos@gmail.com>