ansible/libvirt: Manage the state of the hosts
Use the `state.ini` file to declare a host as "present" or "absent".

When a host is in "present" state, Ansible runs all the necessary tasks
to define/create the host using libvirt, configure the PXE Server.
create the disk, and boot the machine from the network.

When a host is in "absent" state, Ansible runs all the necessary tasks
to destroy the host, configure the PXE Server again, remove all DCHP
leases, and delete the disk.

Signed-off-by: Dimitris Poulopoulos <dimitris.a.poulopoulos@gmail.com>