ansible: Remove Kubespray group variables
Kubespray group variables set the default values of the for the
Kubespray project. It is not a good idea to copy them to the base
project, as newer version may alter them or introduce new ones.

Create a Python script that generates symlinks to the Kubespray
variables, and, if needed, override the default group variables in the
project's group variables, which take precedence.

Signed-off-by: Dimitris Poulopoulos <dimitris.a.poulopoulos@gmail.com>